# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a specialized server for managing trading strategies and margins.",
        "title": "Trades Management API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/user/theme": {
            "patch": {
                "description": "Updates preference (LIGHT/DARK) for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User Theme",
                "parameters": [
                    {
                        "description": "Theme Preference",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateThemeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/user/username": {
            "patch": {
                "description": "Updates the username and invalidates the auth cache",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update Username",
                "parameters": [
                    {
                        "description": "Target Email and New Username",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Update successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "model.UpdateThemeRequest": {
            "type": "object",
            "required": [
                "theme"
            ],
            "properties": {
                "theme": {
                    "enum": [
                        "LIGHT",
                        "DARK"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.UserTheme"
                        }
                    ],
                    "example": "DARK"
                }
            }
        },
        "model.UserDto": {
            "type": "object",
            "required": [
                "confirmPassword",
                "email"
            ],
            "properties": {
                "confirmPassword": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "mobile": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/model.UserRole"
                },
                "theme": {
                    "$ref": "#/definitions/model.UserTheme"
                },
                "userId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.UserRole": {
            "description": "ADMIN or USER access level",
            "type": "string",
            "enum": [
                "ADMIN",
                "USER"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleUser"
            ]
        },
        "model.UserTheme": {
            "description": "LIGHT or DARK theme mode",
            "type": "string",
            "enum": [
                "LIGHT",
                "DARK"
            ],
            "x-enum-varnames": [
                "ThemeLight",
                "ThemeDark"
            ]
        },
...
    }
}